@@ -8,8 +8,11 @@ import org.gradle.api.plugins.JavaPluginExtension
 import org.gradle.kotlin.dsl.configure
 import org.gradle.kotlin.dsl.provideDelegate
 import org.gradle.kotlin.dsl.withType
+import org.jetbrains.kotlin.gradle.dsl.JvmTarget
+import org.jetbrains.kotlin.gradle.dsl.KotlinJvmCompilerOptions
 import org.jetbrains.kotlin.gradle.dsl.KotlinJvmOptions
 import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
+import org.jetbrains.kotlin.gradle.tasks.KotlinJvmCompile
 
 internal fun Project.configureKotlinAndroid(
     commonExtension: CommonExtension<*, *, *, *, *, *>
@@ -26,8 +29,8 @@ internal fun Project.configureKotlinAndroid(
             targetCompatibility = JavaVersion.VERSION_17
 
         }
-        kotlinOptions {
-            jvmTarget = JavaVersion.VERSION_17.toString()
+        kotlinCompilerOptions {
+            jvmTarget.set(JvmTarget.JVM_17)
         }
     }
 }
@@ -42,15 +45,15 @@ internal fun Project.configureKotlinJvm() {
 
 private fun Project.configureJvm() {
     tasks.withType<KotlinCompile>().configureEach {
-        kotlinOptions {
-            jvmTarget = JavaVersion.VERSION_17.toString()
+        compilerOptions {
+            jvmTarget.set(JvmTarget.JVM_17)
             val warningsAsErrors: String? by project
-            allWarningsAsErrors = warningsAsErrors.toBoolean()
-            freeCompilerArgs = listOf("-Xjsr305=strict")
+            allWarningsAsErrors.set(warningsAsErrors.toBoolean())
+            freeCompilerArgs.set(listOf("-Xjsr305=strict"))
         }
     }
 }
 
-fun CommonExtension<*, *, *, *, *, *>.kotlinOptions(block: KotlinJvmOptions.()-> Unit) {
-    (this as ExtensionAware).extensions.configure("kotlinOptions", block)
+fun CommonExtension<*, *, *, *, *, *>.kotlinCompilerOptions(block: KotlinJvmCompilerOptions.() -> Unit) {
+    (this as ExtensionAware).extensions.configure("kotlinCompilerOptions", block)
 }
